+++ pkg/agent/qrm-plugins/mb/controller/policy/crossdomain/global_mb_policy.go
@@ -1,4 +1,4 @@
-package policy
+package crossdomain
 
 import (
 	"fmt"
@@ -9,6 +9,7 @@ import (
 
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/mbdomain"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/mbdomain/quotasourcing"
+	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/config"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/plan"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/strategy"
@@ -340,7 +341,7 @@ func (g *globalMBPolicy) hasHighQoSMB(mbQoSGroups map[qosgroup.QoSGroup]*stat.MB
 
 func NewGlobalMBPolicy(ccdMBMin int, domainManager *mbdomain.MBDomainManager, throttleType, easeType strategy.LowPrioPlannerType,
 	sourcerType string,
-) (DomainMBPolicy, error) {
+) (policy.DomainMBPolicy, error) {
 	domainLeafQuotas := make(map[int]int)
 	for domain := range domainManager.Domains {
 		domainLeafQuotas[domain] = -1
