+++ pkg/agent/qrm-plugins/mb/controller/policy/crossdomain/global_mb_policy_test.go
@@ -1,4 +1,4 @@
-package policy
+package crossdomain
 
 import (
 	"reflect"
